fix loss of data connectivity on devices that don't register their gateways in framework
some devices (namely sholes) do not use the android framework to define a default gateway.
this causes problems when the connection is put into suspension as the default gateway is
removed and never restored.

this solution caches the current gateway (determined via native call) when the device
is removing it due to suspension and the default gateway is not known to framework.
the cached gateway is then used when adding a default gateway in the event that one was
not provided in the interim.

Change-Id: I4a8a5d2e567d11689d7f70118582a92c5ee9b0c5

Conflicts:
	core/java/android/net/NetworkStateTracker.java