telephony: LGEStar: Fix init sequence for p999
The p999's vendor-ril has different request IDs (for the same requests)
than the p990, which was messing up the initialization.
This introduces the correct device-specific IDs, and reorders the
requests, resulting in 2 direct improvements: The battery calls are
now running correctly, and the dual-channel radio is operational,
effectively doubling the attainable bitrate.

Change-Id: I204648595319bae3e36a73a78f475b6dff3b67a7