samsung ril: fix 2g/3g toggle while data call is running
this fixes a problem in the samsung modem code where the
modem is unable to switch the preferred networktype while
a data call is running. The modem might sent a wrong radio
type when the network type is switched during a datacall.
That caused some phones to crash cause they thought they
were CDMA instead of GSM.

This patch basicly does:

check if data connection is up -> disable dataconnection if up ->
wait for dataconnection down -> send request for network type ->
wait for result -> enable dataconnection again -> inform caller
that network type was switched

Change-Id: I940d08a0fcba87d741ebd62ebb00a93039ff8540