Fix race condition in notification swipe handling.
Previously, the code was written under the assumption that ACTION_CANCEL
events are always delivered additionally to the normal (down, move, up)
events. It turned out that this isn't true, so we need to fix the
condition where an ACTION_CANCEL event is delivered right after giving
control to the dispatcher.
Symptom of the previous problem was the swiped notification occasionally
getting stuck, needing a restart of the SystemUI process to get going
again.